Added certain extra features and modified the API

+++ resources/item.py
@@ -1,5 +1,9 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import(
+jwt_required,
+get_jwt_claims,
+fresh_jwt_required
+)
 from models.item import ItemModel
 
 
@@ -16,13 +20,14 @@ class Item(Resource):
                         help="Every item needs a store_id."
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -38,12 +43,18 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin Privileges required bruh.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
             return {'message': 'Item deleted.'}
         return {'message': 'Item not found.'}, 404
+
 
     def put(self, name):
         data = Item.parser.parse_args()
@@ -62,4 +73,4 @@ class Item(Resource):
 
 class ItemList(Resource):
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        return {'items': [x.json() for x in ItemModel.find_all()]}
